rename webdrive related tests in android
